Refactoring (bpf-profile #7)
