+++ src/gen/tests/mod.rs
@@ -2,7 +2,7 @@
 
 mod mock;
 
-use crate::gen::{dump, trace};
+use crate::gen::{resolver, trace};
 use std::fs;
 use std::io::Cursor;
 use std::path::Path;
@@ -25,9 +25,9 @@ fn header_ok() {
 
 #[test]
 fn generate_integral() {
-    let dump = dump::Resolver::default();
+    let resolv = resolver::Resolver::default();
     let reader = Cursor::new(mock::SIMPLE_INPUT);
-    let prof = trace::Profile::new(dump, None);
+    let prof = trace::Profile::new(resolv, None);
     assert!(prof.is_ok());
 
     let mut prof = prof.unwrap();
@@ -42,17 +42,17 @@ fn generate_integral() {
     //println!("{}", std::str::from_utf8(&output).unwrap());
     //=======================================================
 
-    assert_eq!(output.len(), 301);
+    assert_eq!(output.len(), 317);
     assert_eq!(output, mock::SIMPLE_CALLGRIND_INTEGRAL);
 }
 
 #[test]
 fn generate_line_by_line() {
-    let dump = dump::Resolver::default();
+    let resolv = resolver::Resolver::default();
     let reader = Cursor::new(mock::SIMPLE_INPUT);
     let asm_name = "/tmp/generate_line_by_line.asm".to_owned();
     let asm = Path::new(&asm_name);
-    let prof = trace::Profile::new(dump, Some(&asm));
+    let prof = trace::Profile::new(resolv, Some(&asm));
     assert!(prof.is_ok());
 
     let mut prof = prof.unwrap();
@@ -67,7 +67,7 @@ fn generate_line_by_line() {
     //println!("{}", std::str::from_utf8(&output).unwrap());
     //=======================================================
 
-    assert_eq!(output.len(), 488);
+    assert_eq!(output.len(), 504);
     assert_eq!(output, mock::SIMPLE_CALLGRIND_LINE_BY_LINE);
 
     let asm = fs::read_to_string(&asm).unwrap();
